pyproject.toml
[project]
name = "docchat-ai"
version = "0.0.1"
description = "Intelligent Document Q&A System with AI-powered document analysis and conversation"
authors = [
    {name = "DocChat AI Team", email = "team@docchat.ai"}
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["ai", "document", "qa", "chat", "pdf", "langchain", "fastapi", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities"
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.37.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-chroma>=0.0.5",
    "langchain-text-splitters>=0.0.1",
    "python-multipart>=0.0.6",
    "pypdf>=3.17.0",
    "python-dotenv>=1.0.0",
    "watchfiles>=0.21.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[project.urls]
Homepage = "https://github.com/your-username/docchat-ai"
Documentation = "https://docchat-ai.readthedocs.io"
Repository = "https://github.com/your-username/docchat-ai"
Issues = "https://github.com/your-username/docchat-ai/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"